+++ src/database/initialization/has_terminology_set_permission.sql
@@ -1,0 +1,66 @@
+SELECT drop_function('has_terminology_set_permission');
+
+CREATE FUNCTION has_terminology_set_permission(
+	user_keycloak_user_id uuid,
+	user_is_admin boolean,
+	terminology_set_id int,
+	verb permission_grant_verb_t,
+	scope permission_grant_entity_type_t
+) RETURNS boolean AS $$
+DECLARE
+	has_permission boolean;
+BEGIN
+	-- If the user is an administrator, they have all permissions
+	IF user_is_admin THEN
+		RETURN TRUE;
+	END IF;
+
+	-- Check if the user has the specified verb on the specified terminology
+	-- set via direct user grant, group membership, or inherited from the
+	-- owning funder.
+	SELECT EXISTS (
+		SELECT 1
+		FROM terminology_sets ts
+		INNER JOIN permission_grants pg ON (
+			-- Direct terminology set grant
+			(
+				pg.context_entity_type = 'terminologySet'
+				AND pg.terminology_set_id = ts.id
+			)
+			-- Inherited from funder
+			OR (
+				pg.context_entity_type = 'funder'
+				AND pg.funder_short_code = ts.funder_short_code
+			)
+		)
+		WHERE ts.id = has_terminology_set_permission.terminology_set_id
+			AND verb_set_permits_verb(
+				pg.verbs, has_terminology_set_permission.verb
+			)
+			AND scope_set_permits_scope(
+				pg.scope, has_terminology_set_permission.scope
+			)
+			AND (
+				(
+					pg.grantee_type = 'user'
+					AND pg.grantee_user_keycloak_user_id
+						= has_terminology_set_permission.user_keycloak_user_id
+				)
+				OR (
+					pg.grantee_type = 'userGroup'
+					AND EXISTS (
+						SELECT 1
+						FROM ephemeral_user_group_associations euga
+						WHERE euga.user_keycloak_user_id
+							= has_terminology_set_permission.user_keycloak_user_id
+							AND euga.user_group_keycloak_organization_id
+								= pg.grantee_keycloak_organization_id
+							AND NOT is_expired(euga.not_after)
+					)
+				)
+			)
+	) INTO has_permission;
+
+	RETURN has_permission;
+END;
+$$ LANGUAGE plpgsql;